require explicitly including read_leaf_frame bpf code as it does not work with all programs.
Summary:
The context object of some bpf programs (such as scheduler switches) is not compatible with functionality required for reading the leaf frame. Specifically
```
  void* pt1 = (void*)PT_REGS_PARM1(ctx);
  void* pt2 = (void*)PT_REGS_PARM2(ctx);
```

As this is not needed or used in the applications of the oss libs at this point, disable it by default.
It can be re-enabled if needed using a -D flag.

Reviewed By: peremyach

Differential Revision: D77029344

fbshipit-source-id: 59bed36c1bbca71fc1e602f8b5a333dee468b223