Fix bug in readFile implementation
Summary:
Noticed a bug in readFile.
The /proc filesystem is special, and the contents are largely virtual files. This results in some typical filesystem APIs not functioning as usual (such as seekg).
The readFile method is meant to handle this by assuming a file size and reading in chunks.

A bug in the initial implementation caused all reads after the first chunk to fail, as after reading the first chunk, we set readSize = fileSize (renamed chunkSize), then read fileSize - readSize bytes, where fileSize - readSize == 0.

Reviewed By: sonntex

Differential Revision: D82216036

fbshipit-source-id: 5ac999821ce67f4646c423e8aea68904ad3640c1